fixes from copilot
