+++ Resources/Locale/en-US/_CE/zLevel/zLevels.ftl
@@ -1,3 +1,6 @@
 ce-zlevel-falling-popup = {$name} falls from above.
 
 ce-zlevel-look-up-fail = There is a ceiling above you. You cannot look up.
+
+ce-zlevel-voice-from-up = {$name} (from above)
+ce-zlevel-voice-from-down = {$name} (from below)
