+++ Resources/Locale/ru-RU/_CE/zLevel/zLevels.ftl
@@ -1,3 +1,6 @@
 ce-zlevel-falling-popup = {$name} падает сверху.
 
 ce-zlevel-look-up-fail = Над вами потолок. Вы не можете смотреть вверх.
+
+ce-zlevel-voice-from-up = {$name} (сверху)
+ce-zlevel-voice-from-down = {$name} (снизу)
